Move 'is' functions to lib/information
