+++ tests/lib/test_information.py
@@ -1,0 +1,140 @@
+# -*- coding: UTF-8 -*-
+#
+# Copyright 2011-2021 by Dirk Gorissen, Stephen Rauch and Contributors
+# All rights reserved.
+# This file is part of the Pycel Library, Licensed under GPLv3 (the 'License')
+# You may not use this work except in compliance with the License.
+# You may obtain a copy of the Licence at:
+#   https://www.gnu.org/licenses/gpl-3.0.en.html
+
+import pytest
+
+import pycel.excellib
+from pycel.excelutil import (
+    DIV0,
+    NA_ERROR,
+    NUM_ERROR,
+    REF_ERROR,
+    VALUE_ERROR,
+)
+from pycel.lib.function_helpers import load_to_test_module
+from pycel.lib.information import (
+    iserr,
+    iserror,
+    iseven,
+    isna,
+    isnumber,
+    isodd,
+    istext,
+)
+
+
+# dynamic load the lib functions from excellib and apply metadata
+load_to_test_module(pycel.lib.information, __name__)
+
+
+@pytest.mark.parametrize(
+    'value, expected', (
+        (0, False),
+        (1, False),
+        (1.0, False),
+        (-1, False),
+        ('a', False),
+        (((1, NUM_ERROR), ('2', DIV0)), ((False, True), (False, True))),
+        (NA_ERROR, False),
+        (NUM_ERROR, True),
+        (REF_ERROR, True),
+    )
+)
+def test_iserr(value, expected):
+    assert iserr(value) == expected
+
+
+@pytest.mark.parametrize(
+    '_iseven, _isodd, value', (
+        (True, False, -100.1),
+        (True, False, '-100.1'),
+        (True, False, -100),
+        (False, True, -99.9),
+        (True, False, 0),
+        (False, True, 1),
+        (True, False, 0.1),
+        (True, False, '0.1'),
+        (True, False, '2'),
+        (True, False, 2.9),
+        (False, True, 3),
+        (False, True, 3.1),
+        (VALUE_ERROR, VALUE_ERROR, True),
+        (VALUE_ERROR, VALUE_ERROR, False),
+        (VALUE_ERROR, ) * 2 + ('xyzzy', ),
+        (VALUE_ERROR, ) * 3,
+        (DIV0, ) * 3,
+    )
+)
+def test_is_even_odd(_iseven, _isodd, value):
+    assert iseven(value) == _iseven
+    assert isodd(value) == _isodd
+
+
+@pytest.mark.parametrize(
+    'value, expected', (
+        (0, False),
+        (1, False),
+        (1.0, False),
+        (-1, False),
+        ('a', False),
+        (((1, NA_ERROR), ('2', DIV0)), ((False, True), (False, True))),
+        (NUM_ERROR, True),
+        (REF_ERROR, True),
+    )
+)
+def test_iserror(value, expected):
+    assert iserror(value) == expected
+
+
+@pytest.mark.parametrize(
+    'value, expected', (
+        (0, False),
+        (1, False),
+        (1.0, False),
+        (-1, False),
+        ('a', False),
+        (((1, NA_ERROR), ('2', 3)), ((False, True), (False, False))),
+        (NA_ERROR, True),
+        (VALUE_ERROR, False),
+    )
+)
+def test_isna(value, expected):
+    assert isna(value) == expected
+
+
+@pytest.mark.parametrize(
+    'value, expected', (
+        (0, True),
+        (1, True),
+        (1.0, True),
+        (-1, True),
+        ('a', False),
+        (((1, NA_ERROR), ('2', 3)), ((True, False), (False, True))),
+        (NA_ERROR, False),
+        (VALUE_ERROR, False),
+    )
+)
+def test_isnumber(value, expected):
+    assert isnumber(value) == expected
+
+
+@pytest.mark.parametrize(
+    'value, expected', (
+        ('a', True),
+        (1, False),
+        (1.0, False),
+        (None, False),
+        (DIV0, False),
+        (((1, NA_ERROR), ('2', 3)), ((False, False), (True, False))),
+        (NA_ERROR, False),
+        (VALUE_ERROR, False),
+    )
+)
+def test_istext(value, expected):
+    assert istext(value) == expected
